refactor(frontend): 사이드바 UI를 layout으로 분리
페이지 공통 사이드바를 layout.tsx로 이동하여
하위 라우트에서 공유할 수 있도록 구조 변경
